Create Api Controller
In this assignment, you are required to develop a simple ASP.NET Core Web API project. The project will manage a list of products, and an API will be created to perform basic CRUD (Create, Read, Update, Delete) operations on this product list. The required CRUD operations will be performed through Swagger.